fix(context): carry the pipeline score onto every bundled memory row
`api::context::run` kept only the ids of the pipeline's hits, so
`bundle::assemble` had no score to carry and every memory row in a
context bundle reported `score: 0.0` on both the CLI and
`GET|POST /api/v1/context`. A consumer could not tell the top-ranked
memory from the last one.

- `bundle::assemble` takes `&[RankedMemory]` (id + score) instead of bare
  ids; `RankedMemory: From<Reranked>` carries `parts.final_score`.
- `MemoryBundleRow::score` is `f64`, the pipeline's native type, so the
  bundle reports the same number `search --json` gives the same query.
- `tty::score` takes `f64`, dropping the `as f32` casts in the `search`
  and `search-code` renderers.

Closes #152.

+++ src/output/search.rs
@@ -235,7 +235,7 @@ pub fn write_tty<S: ::std::hash::BuildHasher>(
         writeln!(
             out,
             "{}  {}  {}{}{}",
-            tty::score(hit.parts.final_score as f32),
+            tty::score(hit.parts.final_score),
             source_label(hit.source),
             hit.memory_id,
             suffix,
